requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.3.0
aiohappyeyeballs=2.6.1=pyhd8ed1ab_0
aiohttp=3.11.18=py311h4921393_0
aiosignal=1.3.2=pyhd8ed1ab_0
appnope=0.1.4=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
attrs=25.3.0=pyh71513ae_0
brotli=1.1.0=hd74edd7_2
brotli-bin=1.1.0=hd74edd7_2
brotli-python=1.1.0=py311h3f08180_2
bzip2=1.0.8=h99b78c6_7
ca-certificates=2025.4.26=hbd8a1cb_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py311h3a79f62_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.2=py311h210dab8_0
coverage=7.8.0=py311h4921393_0
cpython=3.11.12=py311hd8ed1ab_0
cycler=0.12.1=pyhd8ed1ab_1
debugpy=1.8.14=py311h155a34a_0
decorator=5.2.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.2.0=pyhd8ed1ab_0
filelock=3.18.0=pyhd8ed1ab_0
fonttools=4.57.0=py311h4921393_0
freetype=2.13.3=hce30654_1
frozenlist=1.5.0=py311h4921393_1
fsspec=2025.3.2=pyhd8ed1ab_0
gmp=6.3.0=h7bae524_2
gmpy2=2.2.1=py311hb5d9ff4_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.6.1=pyha770c72_0
iniconfig=2.0.0=pyhd8ed1ab_1
ipykernel=6.29.5=pyh57ce528_0
ipython=9.2.0=pyhfb0248b_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pyhd8ed1ab_0
joblib=1.4.2=pyhd8ed1ab_1
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.7=py311h2c37856_0
krb5=1.21.3=h237132a_0
lcms2=2.17=h7eeda09_0
lerc=4.0.0=hd64df32_1
libabseil=20250127.1=cxx17_h07bc746_0
libblas=3.9.0=31_h10e41b3_openblas
libbrotlicommon=1.1.0=hd74edd7_2
libbrotlidec=1.1.0=hd74edd7_2
libbrotlienc=1.1.0=hd74edd7_2
libcblas=3.9.0=31_hb3479ef_openblas
libcxx=20.1.4=ha82da77_0
libdeflate=1.23=h5773f1b_0
libedit=3.1.20250104=pl5321hafb1f1b_0
libexpat=2.7.0=h286801f_0
libffi=3.4.6=h1da3d7d_1
libfreetype=2.13.3=hce30654_1
libfreetype6=2.13.3=h1d14073_1
libgfortran=14.2.0=heb5dd2a_105
libgfortran5=14.2.0=h2c44a93_105
libjpeg-turbo=3.1.0=h5505292_0
liblapack=3.9.0=31_hc9a63f6_openblas
liblzma=5.8.1=h39f12f2_0
libmpdec=4.0.0=h99b78c6_0
libopenblas=0.3.29=openmp_hf332438_0
libpng=1.6.47=h3783ad8_0
libprotobuf=5.29.3=hccd9074_1
libsodium=1.0.20=h99b78c6_0
libsqlite=3.49.1=h3f77e49_2
libtiff=4.7.0=h551f018_4
libtorch=2.6.0=cpu_generic_h9a9d006_4
libuv=1.50.0=h5505292_0
libwebp-base=1.5.0=h2471fea_0
libxcb=1.17.0=hdb1d25a_0
libzlib=1.3.1=h8359307_2
littleutils=0.2.4=pyhd8ed1ab_1
llvm-openmp=20.1.3=hdb05f8b_0
markupsafe=3.0.2=py311h4921393_1
matplotlib=3.10.1=py311ha1ab1f8_0
matplotlib-base=3.10.1=py311h031da69_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mpc=1.3.1=h8f1351a_1
mpfr=4.2.1=hb693164_3
mpmath=1.3.0=pyhd8ed1ab_1
multidict=6.4.3=py311h30e7462_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.5=h5e97a16_3
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pyh267e887_2
nomkl=1.0=h5ca1d4c_0
numpy=1.26.4=py311h7125741_0
ogb=1.3.6=pyhd8ed1ab_1
openjpeg=2.5.3=h8a3d83b_0
openssl=3.5.0=h81ee809_0
optree=0.15.0=py311h210dab8_0
outdated=0.2.2=pyhd8ed1ab_2
packaging=25.0=pyh29332c3_1
pandas=2.2.3=py311hca32420_3
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=py311hb9ba9e9_0
pip=25.1=pyh8b19718_0
platformdirs=4.3.7=pyh29332c3_0
pluggy=1.5.0=pyhd8ed1ab_1
prompt-toolkit=3.0.51=pyha770c72_0
propcache=0.3.1=py311h4921393_0
psutil=7.0.0=py311h917b07b_0
pthread-stubs=0.4=hd74edd7_1002
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pybind11=2.13.6=pyh1ec8472_2
pybind11-global=2.13.6=pyh415d2e4_2
pycparser=2.22=pyh29332c3_1
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.3=pyhd8ed1ab_1
pysocks=1.7.1=pyha55dd90_7
pytest=8.3.5=pyhd8ed1ab_0
pytest-cov=6.1.1=pyhd8ed1ab_0
python=3.11.12=hc22306f_0_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-tzdata=2025.2=pyhd8ed1ab_0
python-wget=3.2=py_0
python_abi=3.11=7_cp311
pytorch=2.6.0=cpu_generic_py311_hc7d8f6d_4
pytorch_cluster=1.6.3=py311hfb428f7_7
pytorch_geometric=2.6.1=pyhecae5ae_2
pytorch_scatter=2.1.2=cpu_py311hfb428f7_12
pytorch_sparse=0.6.18=py311hdb4de42_7
pytorch_spline_conv=1.2.2=cpu_py311hfb428f7_8
pytz=2025.2=pyhd8ed1ab_0
pyzmq=26.4.0=py311h01f2145_0
qhull=2020.2=h420ef59_5
readline=8.2=h1d1bf99_2
requests=2.32.3=pyhd8ed1ab_1
scikit-learn=1.6.1=py311h47fa2fb_0
scipy=1.15.2=py311h0675101_0
setuptools=75.8.2=pyhff2d567_0
six=1.17.0=pyhd8ed1ab_0
sleef=3.8=h8391f65_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.14.0=pyh2585a3b_105
threadpoolctl=3.6.0=pyhecae5ae_0
tk=8.6.13=h5083fa2_1
toml=0.10.2=pyhd8ed1ab_1
tomli=2.2.1=pyhd8ed1ab_1
torch-geometric=2.6.1=pyhd8ed1ab_0
tornado=6.4.2=py311h917b07b_0
tqdm=4.67.1=pyhd8ed1ab_1
traitlets=5.14.3=pyhd8ed1ab_1
typing-extensions=4.13.2=h0e9735f_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025b=h78e105d_0
unicodedata2=16.0.0=py311h917b07b_0
urllib3=2.4.0=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=pyhd8ed1ab_1
xorg-libxau=1.0.12=h5505292_0
xorg-libxdmcp=1.1.5=hd74edd7_0
yarl=1.20.0=py311h4921393_0
zeromq=4.3.5=hc1bb282_7
zipp=3.21.0=pyhd8ed1ab_1
zstandard=0.23.0=py311h917b07b_2
zstd=1.5.7=h6491c7d_2
